requirements.txt
requests
beautifulsoup4
langchain
langchain-text-splitters
langchain-community
langchain-google-genai
unstructured
faiss-cpu
sentence-transformers
rank_bm25
tiktoken
streamlit
transformers
pypdf
ipykernel
python-dotenv
